+++ agent-watchdog/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Added
+- Incremented version for build.
+
 ## 1.0.1 - 2025-05-12
 ### Changed
 - Updated the Windows build script to use `pushd` and `popd` when running from a UNC path.